Configuration struct
Start of implementation of a single configuration struct
to configure many machine parameters at once. Design of this object
is in flux at the moment - possible best design will avoid coupling
with the Machine object and only use public member functions to
configure the machine object.

If using that design it would be preferable to delay the emission of the
preamble until after configuration is complete to avoid e.g. preamble
with metric and then instant switch to imperial.